// valorant-api/models/Affinities.ts
/**
 * 
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 3.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export enum Affinities {
    Eu = 'eu',
    Na = 'na',
    Latam = 'latam',
    Br = 'br',
    Ap = 'ap',
    Kr = 'kr'
}