// valorant-api/models/V1LifetimeMatches.ts
/**
 * 
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 3.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { V1LifetimeMatchesItem } from '../models/V1LifetimeMatchesItem';
import { V1LifetimeMatchesResults } from '../models/V1LifetimeMatchesResults';
import { HttpFile } from '../http/http';

export class V1LifetimeMatches {
    'status'?: number;
    'name'?: string;
    'tag'?: string;
    'results'?: V1LifetimeMatchesResults;
    'data'?: Array<V1LifetimeMatchesItem>;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "status",
            "baseName": "status",
            "type": "number",
            "format": ""
        },
        {
            "name": "name",
            "baseName": "name",
            "type": "string",
            "format": ""
        },
        {
            "name": "tag",
            "baseName": "tag",
            "type": "string",
            "format": ""
        },
        {
            "name": "results",
            "baseName": "results",
            "type": "V1LifetimeMatchesResults",
            "format": ""
        },
        {
            "name": "data",
            "baseName": "data",
            "type": "Array<V1LifetimeMatchesItem>",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return V1LifetimeMatches.attributeTypeMap;
    }

    public constructor() {
    }
}